refactor: change user model file name
