--- src/__tests__/index.test.ts
@@ -1,5 +1,0 @@
-describe('Test sample', () => {
-  it('1 + 1 equals 2', () => {
-    expect(1 + 1).toEqual(2)
-  })
-})
