+++ src/use_cases/__tests__/login.test.ts
@@ -1,5 +1,5 @@
 import login from '../login'
-import User from '../../models/User'
+import UserModel from '../../models/UserModel'
 import { encryptPassword } from '../../utils/crypt'
 import { connect, disconnect } from '../../database'
 
@@ -13,14 +13,14 @@ describe('Use Case: Login', () => {
   })
 
   beforeEach(async () => {
-    await User.deleteMany({})
+    await UserModel.deleteMany({})
   })
 
   describe('when passing a valid email and password', () => {
     it('returns the logged in user', async () => {
       const email = 'john@test.com'
       const password = '123456'
-      const user = await createUser({ email, password })
+      const user = await createUser(email, password)
 
       const loggedInUser = await login(email, password)
 
@@ -28,8 +28,8 @@ describe('Use Case: Login', () => {
     })
   })
 
-  const createUser = async ({ email, password }) => {
-    const user = new User({
+  const createUser = async (email: string, password: string) => {
+    const user = new UserModel({
       username: 'John',
       email,
       password: encryptPassword(password)
